dbOut Dual Channel Part 1

+++ TestTapeGenerator/TapeInfo.h
@@ -8,11 +8,11 @@ const char* TAPELIST_VERSION = "TapeList: " __DATE__ " " __TIME__;
 class RecordStep
 {
 public:
-    int Frequency;
-    int Time;
-    int Level;
+    uint32_t Frequency;
+    uint8_t Time;
+    int8_t Level;
     const __FlashStringHelper* Comment;
-    RecordStep(int Frequency_, int Time_, int Level_, const __FlashStringHelper* Comment_ = 0) : Frequency(Frequency_), Time(Time_), Level(Level_), Comment(Comment_)
+    RecordStep(uint32_t Frequency_, uint8_t Time_, int8_t Level_, const __FlashStringHelper* Comment_ = 0) : Frequency(Frequency_), Time(Time_), Level(Level_), Comment(Comment_)
     {
     }
 
@@ -48,19 +48,19 @@ class TapeInfo
 public:
     enum TapeFormat { Cassette, Reel };
     const __FlashStringHelper* Description;
-    uint16_t Tracks;
-    uint16_t Length;
+    uint8_t Tracks;
+    uint8_t Length;
     uint16_t Flux;
     TapeFormat Format;
-    double Target;
+    int8_t Target;
     std::vector<RecordStep*> RecordSteps;
 
     TapeInfo(
         const __FlashStringHelper* Description_,
-        uint16_t Tracks_,
+        uint8_t Tracks_,
         uint16_t Flux_,
         TapeFormat Format_,
-        double Target_,
+        int8_t Target_,
         std::vector<RecordStep*> RecordSteps_) : Description(Description_), Tracks(Tracks_), Flux(Flux_), Format(Format_), Target(Target_), RecordSteps(RecordSteps_)
     {
         Length = 0;
@@ -125,14 +125,14 @@ public:
     static TapeInfo* Get(Tapes tape) {
         switch (tape) {
         case AKAI_GX_75_95_TEST_TAPE:
-            return new TapeInfo(F("AKAI GX-75/95 Test Tape"), 2, 200, TapeInfo::Cassette, 0.0, {
+            return new TapeInfo(F("AKAI GX-75/95 Test Tape"), 2, 200, TapeInfo::Cassette, 0, {
                 new RecordStep(315, 180, 0, F("Level")),
                 new RecordStep(1000, 180, 0, F("Azimuth")),
                 new RecordStep(3150, 180, 0),
                 new RecordStep(10000, 180, -15)
                 });
         case NAKAMICHI_TEST_TAPE:
-            return new TapeInfo(F("Nakamichi Test Tape"), 2, 200, TapeInfo::Cassette, 0.0, {
+            return new TapeInfo(F("Nakamichi Test Tape"), 2, 200, TapeInfo::Cassette, 0, {
                 new RecordStep(20000, 180, -20, F("DA09001A 20 kHz Frequency Response")),
                 new RecordStep(15000, 180, -20, F("DA09002A 15 kHz Frequency Response")),
                 new RecordStep(10000, 180, -20, F("DA09003A 10 kHz Frequency Response")),
@@ -141,7 +141,7 @@ public:
                 new RecordStep(3000, 180, 0, F("DA09006A 3 kHz Speed and Wow & Flutter"))
                 });
         case PANASONIC_QZZCFM_TEST_TAPE:
-            return new TapeInfo(F("Panasonic QZZCFM Test Tape"), 2, 200, TapeInfo::Cassette, 0.0, {
+            return new TapeInfo(F("Panasonic QZZCFM Test Tape"), 2, 200, TapeInfo::Cassette, 0, {
                 new RecordStep(315, 180, -2, F("Level")),
                 new RecordStep(3000, 180, -10, F("Azimuth")),
                 new RecordStep(8000, 180, -20),
@@ -155,7 +155,7 @@ public:
                 new RecordStep(12500, 20, -20)
                 });
         case PLAYBACK_EQ_TEST_TAPE:
-            return new TapeInfo(F("Playback EQ Test Tape"), 2, 200, TapeInfo::Cassette, 0.0, {
+            return new TapeInfo(F("Playback EQ Test Tape"), 2, 200, TapeInfo::Cassette, 0, {
                 //new RecordStep(1000, 120, 0, "1 kHz Reference Level"),
                 //new RecordStep(10000, 5, -20, "Alternating 10kHz"),
                 //new RecordStep(1000, 5, -20, "Alternating 1kHz"),
@@ -208,7 +208,7 @@ public:
                 new RecordStep(10000, 120, -20, F("Azimuth"))
                 });
         case REVOX_B215_TEST_TAPE:
-            return   new TapeInfo(F("Revox B-215 Test Tape"), 2, 200, TapeInfo::Cassette, 0.0, {
+            return   new TapeInfo(F("Revox B-215 Test Tape"), 2, 200, TapeInfo::Cassette, 0, {
                 new RecordStep(1000, 180, 2),
                 new RecordStep(1000, 180, 0),
                 new RecordStep(3150, 180, 0),
@@ -216,30 +216,30 @@ public:
                 new RecordStep(16000, 180, -20)
                 });
         case REVOX_B77_TEST_TAPE_1:
-            return new TapeInfo(F("Revox B-77 Test Tape #1"), 2, 257, TapeInfo::Reel, 0.0, {
+            return new TapeInfo(F("Revox B-77 Test Tape #1"), 2, 257, TapeInfo::Reel, 0, {
                 new RecordStep(1000, 120, 0, F("Level")),
                 new RecordStep(1000, 60, -20, F("Azimuth")),
                 new RecordStep(10000, 120, -20, F("Azimuth")),
                 new RecordStep(10000, 120, 0, F("Azimuth"))
                 });
         case REVOX_B77_TEST_TAPE_2:
-            return new TapeInfo(F("Revox B-77 Test Tape #2"), 4, 257, TapeInfo::Reel, 0.0, {
+            return new TapeInfo(F("Revox B-77 Test Tape #2"), 4, 257, TapeInfo::Reel, 0, {
                 new RecordStep(1000, 120, 0, F("Level")),
                 new RecordStep(1000, 60, -20, F("Azimuth")),
                 new RecordStep(10000, 120, -20, F("Azimuth")),
                 new RecordStep(10000, 120, 0, F("Azimuth"))
                 });
         case STUDER_A710_TEST_TAPE:
-            return new TapeInfo(F("Studer A710 Test Tape"), 2, 200, TapeInfo::Cassette, 0.0, {
+            return new TapeInfo(F("Studer A710 Test Tape"), 2, 200, TapeInfo::Cassette, 0, {
                 new RecordStep(315, 460, 0, F("Level")),
                 new RecordStep(10000, 460, -20, F("Azimuth"))
                 });
         case TANDBERG_24_TEST_TAPE:
-            return new TapeInfo(F("Tandberg #24 Test Tape"), 2, 250, TapeInfo::Cassette, 0.0, {
+            return new TapeInfo(F("Tandberg #24 Test Tape"), 2, 250, TapeInfo::Cassette, 0, {
                 new RecordStep(1000, 4800, +2, F("Flux: 250nW/m"))
                 });
         case TEAC_MTT_MULTI_TEST_TAPE:
-            return new TapeInfo(F("Tandberg #24 Test Tape"), 2, 250, TapeInfo::Cassette, 0.0, {
+            return new TapeInfo(F("Tandberg #24 Test Tape"), 2, 250, TapeInfo::Cassette, 0, {
                 new RecordStep(315, 120, 0, F("MTT-212N")),
                 new RecordStep(315, 120, -4, F("MTT-212CN")),
                 new RecordStep(1000, 120, -4, F("MTT-212EN")),
@@ -251,7 +251,7 @@ public:
                 new RecordStep(12500, 120, -24, F("MTT-118NA"))
                 });
         case UNIVERSAL_TEST_TAPE:
-            return new TapeInfo(F("Universal Test Tape"), 2, 200, TapeInfo::Cassette, 0.0, {
+            return new TapeInfo(F("Universal Test Tape"), 2, 200, TapeInfo::Cassette, 0, {
                 new RecordStep(1000, 120, 0, F("Level")),
                 new RecordStep(31.5, 60, 0, F("Azimuth")),
                 new RecordStep(31.5, 60, -3),
@@ -325,22 +325,22 @@ public:
                 new RecordStep(1000, 120, 0)
                 });
         case WOW_AND_FLUTTER_TEST_TAPE_1:
-            return new TapeInfo(F("Wow & Flutter Test Tape #1"), 2, 200, TapeInfo::Cassette, 0.0, {
+            return new TapeInfo(F("Wow & Flutter Test Tape #1"), 2, 200, TapeInfo::Cassette, 0, {
                 new RecordStep(3000, 120, 0, F("DIN Standard")),
                 new RecordStep(3150, 120, 0, F("JIS Standard"))
                 });
         case WOW_AND_FLUTTER_TEST_TAPE_2:
-            return new TapeInfo(F("Wow & Flutter Test Tape #2"), 2, 257, TapeInfo::Reel, 0.0, {
+            return new TapeInfo(F("Wow & Flutter Test Tape #2"), 2, 257, TapeInfo::Reel, 0, {
                 new RecordStep(3000, 120, 0, F("DIN Standard")),
                 new RecordStep(3150, 120, 0, F("JIS Standard"))
                 });
         case WOW_AND_FLUTTER_TEST_TAPE_3:
-            return new TapeInfo(F("Wow & Flutter Test Tape #3"), 4, 185, TapeInfo::Reel, -5.0, {
+            return new TapeInfo(F("Wow & Flutter Test Tape #3"), 4, 185, TapeInfo::Reel, -5, {
                 new RecordStep(3000, 120, 0, F("DIN Standard")),
                 new RecordStep(3150, 120, 0, F("JIS Standard"))
                 });
         case WOW_AND_FLUTTER_TEST_TAPE_4:
-            return new TapeInfo(F("Wow & Flutter Test Tape #4"), 4, 257, TapeInfo::Reel, 0.0, {
+            return new TapeInfo(F("Wow & Flutter Test Tape #4"), 4, 257, TapeInfo::Reel, 0, {
                 new RecordStep(3000, 120, 0, F("DIN Standard")),
                 new RecordStep(3150, 120, 0, F("JIS Standard"))
                 });
